+++ docs/conf.py
@@ -26,7 +26,16 @@ sys.path.insert(0, os.path.dirname(os.path.abspath('.')))
 
 # Add any Sphinx extension module names here, as strings. They can be extensions
 # coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
-extensions = ['sphinx.ext.autodoc', 'sphinx.ext.intersphinx', 'sphinx.ext.viewcode']
+extensions = [
+    'sphinx.ext.autodoc',
+    'sphinx.ext.extlinks',
+    'sphinx.ext.intersphinx',
+    'sphinx.ext.viewcode',
+]
+
+extlinks = {
+    'issue': ('https://github.com/rbarrois/factory_boy/issues/%s', 'issue #'),
+}
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']