+++ docs/changelog.rst
@@ -17,6 +17,8 @@ ChangeLog
       to ease resetting the sequence counter for a given factory.
     - Add debug messages to ``factory`` logger.
     - Add a :meth:`~factory.Iterator.reset` method to :class:`~factory.Iterator` (:issue:`63`)
+    - Add support for the SQLAlchemy ORM through :class:`~factory.alchemy.SQLAlchemyModelFactory`
+      (:issue:`64`, thanks to `Romain Commandé <https://github.com/rcommande>`_)
 
 *Bugfix*