+++ docs/changelog.rst
@@ -19,6 +19,8 @@ ChangeLog
     - Add a :meth:`~factory.Iterator.reset` method to :class:`~factory.Iterator` (:issue:`63`)
     - Add support for the SQLAlchemy ORM through :class:`~factory.alchemy.SQLAlchemyModelFactory`
       (:issue:`64`, thanks to `Romain Commandé <https://github.com/rcommande>`_)
+    - Add :class:`factory.django.FileField` and :class:`factory.django.ImageField` hooks for
+      related Django model fields (:issue:`52`)
 
 *Bugfix*